# Avoid redundant `HasValue` call in lifted custom binary operator