Fixes https://github.com/dotnet/roslyn/issues/56875.

<details>

<summary>

### Microbenchmark ([source](https://github.com/jjonescz/RoslynIssue56875))

</summary>

| Method | Runtime              | Mean      | Ratio | Code Size |
|------- |--------------------- |----------:|------:|----------:|
| Before | .NET 8.0             | 0.3349 ns |  1.00 |      38 B |
| After  | .NET 8.0             | 0.1841 ns |  0.55 |      36 B |
| Before | .NET Framework 4.8.1 | 1.3028 ns |  3.88 |     114 B |
| After  | .NET Framework 4.8.1 | 0.4789 ns |  1.43 |      55 B |

```cs
public bool Before(DateTime? d) => d == DateTime.MaxValue;
public bool After(DateTime? d) => d.HasValue && d.GetValueOrDefault() == DateTime.MaxValue;
```

</details>