@jaredpar 2023-11-27
> It would be good to run a microbenchmark on this (manually writing the old and new lowered version in C# directly) to make sure that we didn't just break the JIT's pattern recognition.

@jkotas, @davidwrighton are micro benchmarks, or asm diffs, sufficient for us to feel confident that we didn't regress a JIT pattern match when we update IL? Or is there a different approach that we should take?