@jkotas 2023-11-27
I am not worried about perf regressions from this change. JIT does not have pattern matching for Nullable.HasValue.

On .NET Framework, you should see an improvement in microbenchmarks.
On .NET 8, you should see roughly same perf before/after in microbenchmarks. The current JIT is able optimize out the redundant HasValue check.

cc @AndyAyersMS