@AndyAyersMS 2023-11-28
> The current JIT is able optimize out the redundant HasValue check.

The .NET 8 improvement is deferring the load of the nullable payload which unblocks more efficient testing of the flag.  The jit  does not do any sort of partial dead code optimization so if you have something like the before case
```
a = x.f1;
b = x.f2;
if (a ..)
{
     ... b; // only use of b
}
```
we won't defer the second load. Also we won't reorder the two loads.

You should also see some improvement for the null case.