@jjonescz 2023-12-04
@dotnet/roslyn-compiler for a second review, thanks